requirements.txt
aiogram==3.22.0
python-dotenv==1.0.0
aiomysql==0.2.0
sqlalchemy==2.0.23
alembic==1.12.1
